+++ Flutter/room_flutter_example/lib/pages/conference_tuikit/create_room/controller.dart
@@ -1,3 +1,4 @@
+import 'dart:math';
 import 'package:get/get.dart';
 import 'package:room_flutter_example/common/index.dart';
 import 'package:rtc_conference_tui_kit/common/index.dart';
@@ -5,9 +6,10 @@ import 'package:rtc_conference_tui_kit/rtc_conference_tuikit.dart';
 import 'package:rtc_room_engine/rtc_room_engine.dart';
 
 class CreateRoomController extends GetxController {
+  var isOperating = false;
   CreateRoomController();
+  final int numberOfDigits = 6;
   RxString roomTypeString = 'freeToSpeakRoom'.tr.obs;
-
   RxString chooseSpeechMode = 'freeToSpeakRoom'.obs;
   var roomSpeechMode = TUISpeechMode.freeToSpeak;
 
@@ -32,29 +34,39 @@ class CreateRoomController extends GetxController {
     Get.back();
   }
 
-  getRoomId() {
-    return (('${UserStore.to.userModel.userId}_room_kit').hashCode & 0x3B9AC9FF)
-        .toString();
+  String _getRoomId() {
+    Random random = Random();
+    int minNumber = pow(10, numberOfDigits - 1).toInt();
+    int maxNumber = pow(10, numberOfDigits).toInt() - 1;
+    int randomNumber = random.nextInt(maxNumber - minNumber) + minNumber;
+    String roomId = randomNumber.toString();
+    return roomId;
   }
 
-  void createRoom() {
+  void createRoom() async {
+    if (isOperating) {
+      return;
+    }
+
     var roomKit = TUIRoomKit.createInstance();
-    TUIRoomInfo roomInfo = TUIRoomInfo(roomId: getRoomId());
+    TUIRoomInfo roomInfo = TUIRoomInfo(roomId: _getRoomId());
     roomInfo.name = UserStore.to.userModel.userName;
     roomInfo.speechMode = roomSpeechMode;
-    roomKit.createRoom(roomInfo).then((value) {
-      if (value.code == TUIError.success) {
-        roomKit
-            .enterRoom(getRoomId(), UserStore.to.openMicrophone.value,
-                UserStore.to.openCamera.value, UserStore.to.userSpeaker.value)
-            .then((value) {
-          if (value.code != TUIError.success) {
-            makeToast(msg: value.message!);
-          }
-        });
-      } else {
-        makeToast(msg: value.message!);
+    isOperating = true;
+    var createRoomResult = await roomKit.createRoom(roomInfo);
+    if (createRoomResult.code == TUIError.success) {
+      var enterRoomResult = await roomKit.enterRoom(
+          roomInfo.roomId,
+          UserStore.to.openMicrophone.value,
+          UserStore.to.openCamera.value,
+          UserStore.to.userSpeaker.value);
+
+      if (enterRoomResult.code != TUIError.success) {
+        makeToast(msg: enterRoomResult.message!);
       }
-    });
+    } else {
+      makeToast(msg: createRoomResult.message!);
+    }
+    isOperating = false;
   }
 }