+++ Flutter/room_flutter_example/lib/pages/conference_tuikit/prepare/controller.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:room_flutter_example/common/index.dart';
 
@@ -14,5 +15,14 @@ class PrepareController extends GetxController {
 
   finishPage() {
     Get.back();
+  }
+
+  switchLanguage() {
+    if (Get.locale!.languageCode ==
+        TranslationService.fallbackLocale.languageCode) {
+      Get.updateLocale(const Locale('zh', 'CN'));
+    } else {
+      Get.updateLocale(TranslationService.fallbackLocale);
+    }
   }
 }