+++ Flutter/room_flutter_example/lib/pages/conference_tuikit/prepare/widgets/create_room.dart
@@ -10,14 +10,17 @@ class CreateRoomWidget extends GetView<PrepareController> {
   @override
   Widget build(BuildContext context) {
     return ElevatedButton(
-      child: Row(mainAxisAlignment: MainAxisAlignment.center, children: [
-        Image.asset(AssetsImages.createRoom, width: 16, height: 16),
-        const SizedBox(width: 5),
-        Text(
-          'createRoom'.tr,
-          style: const TextStyle(fontSize: 16),
-        )
-      ]),
+      child: Row(
+        mainAxisAlignment: MainAxisAlignment.center,
+        children: [
+          Image.asset(AssetsImages.createRoom, width: 16, height: 16),
+          const SizedBox(width: 5),
+          Text(
+            'createRoom'.tr,
+            style: const TextStyle(fontSize: 16, color: Colors.white),
+          )
+        ],
+      ),
       onPressed: () => controller.toCreateRoomPage(),
     );
   }