+++ Flutter/rtc_conference_tui_kit/lib/common/widgets/index.dart
@@ -3,7 +3,6 @@ library widgets;
 export './copy_text_button.dart';
 export './info_list.dart';
 export './drop_down_button.dart';
-export './invite_sheet.dart';
 export './search_bar.dart';
 export './rounded_container.dart';
 export './bottom_sheet.dart';