+++ Flutter/rtc_conference_tui_kit/lib/pages/conference_main/widgets/bottom_view/widgets/bottom_button_item.dart
@@ -48,6 +48,8 @@ class BottomButtonItemWidget extends GetView<BottomViewController> {
                   isSelected.value ? selectedText ?? text : text,
                   textAlign: TextAlign.center,
                   style: RoomTheme.defaultTheme.textTheme.labelSmall,
+                  maxLines: 1,
+                  overflow: TextOverflow.ellipsis,
                 ),
               ),
             ),