+++ Flutter/rtc_conference_tui_kit/lib/pages/conference_main/widgets/user_list/view.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:rtc_conference_tui_kit/common/index.dart';
+import 'package:rtc_conference_tui_kit/pages/conference_main/widgets/widgets.dart';
 import 'package:rtc_room_engine/api/room/tui_room_define.dart';
 
-import 'index.dart';
 import './widgets/widgets.dart';
 
 class UserListWidget extends GetView<UserListController> {