+++ Flutter/rtc_conference_tui_kit/lib/pages/conference_main/widgets/video_seat/video_item/index.dart
@@ -1,3 +1,0 @@
-library video_item;
-
-export './view.dart';