+++ Flutter/rtc_conference_tui_kit/lib/pages/conference_main/widgets/video_seat/video_layout/widgets/with_draggable_window_widget.dart
@@ -47,14 +47,6 @@ class WithDraggableWindowWidget extends GetView<VideoLayoutController> {
 
   @override
   Widget build(BuildContext context) {
-    return GetBuilder<VideoLayoutController>(
-      init: VideoLayoutController(),
-      id: "with_draggable_window",
-      builder: (_) {
-        return SafeArea(
-          child: _buildView(),
-        );
-      },
-    );
+    return _buildView();
   }
 }