+++ Flutter/room_flutter_example/CHANGELOG.md
@@ -1,3 +1,8 @@
+### 1.7.0(2023-12-18)
+
+- Release the official version of room_flutter_example
+- Upgrade the dependencies of rtc_conference_tui_kit to the official version
+
 ### 1.6.1-rc(2023-11-14)
 
 - Release candidate version of room_flutter_example