+++ Flutter/rtc_conference_tui_kit/CHANGELOG.md
@@ -1,4 +1,10 @@
+### 1.7.0(2023-12-18)
+
+- Release the official version of rtc_conference_tui_kit
+- Fixed some known bugs and issues
+- Upgrade the dependencies of rtc_room_engine
+
 ### 1.6.1-rc(2023-11-14)
 
-- Release candidate version of rtc_conference_tuikit
+- Release candidate version of rtc_conference_tui_kit
 - Supports only Android and iOS platforms